+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (Feature) ConfigMap Inspector
 - (Feature) Envoy Image Layer
 - (Feature) Add ArangoDeployment ServerGroupType
+- (Feature) ServerGroup Pointer
 
 ## [1.2.42](https://github.com/arangodb/kube-arangodb/tree/1.2.42) (2024-07-23)
 - (Maintenance) Go 1.22.4 & Kubernetes 1.29.6 libraries